Move asyncio client and server out of the way.
