+++ tests/test_asyncio_client_server.py
@@ -13,7 +13,8 @@ import urllib.error
 import urllib.request
 import warnings
 
-from websockets.client import *
+from websockets.asyncio_client import *
+from websockets.asyncio_server import *
 from websockets.datastructures import Headers
 from websockets.exceptions import (
     ConnectionClosed,
@@ -31,7 +32,6 @@ from websockets.handshake_legacy import build_response
 from websockets.http import USER_AGENT
 from websockets.http_legacy import read_response
 from websockets.protocol import State
-from websockets.server import *
 from websockets.uri import parse_uri
 
 from .test_protocol import MS
@@ -1072,7 +1072,7 @@ class CommonClientServerTests:
         self.run_loop_once()
 
     @with_server()
-    @unittest.mock.patch("websockets.server.read_request")
+    @unittest.mock.patch("websockets.asyncio_server.read_request")
     def test_server_receives_malformed_request(self, _read_request):
         _read_request.side_effect = ValueError("read_request failed")
 
@@ -1080,7 +1080,7 @@ class CommonClientServerTests:
             self.start_client()
 
     @with_server()
-    @unittest.mock.patch("websockets.client.read_response")
+    @unittest.mock.patch("websockets.asyncio_client.read_response")
     def test_client_receives_malformed_response(self, _read_response):
         _read_response.side_effect = ValueError("read_response failed")
 
@@ -1089,7 +1089,7 @@ class CommonClientServerTests:
         self.run_loop_once()
 
     @with_server()
-    @unittest.mock.patch("websockets.client.build_request")
+    @unittest.mock.patch("websockets.asyncio_client.build_request")
     def test_client_sends_invalid_handshake_request(self, _build_request):
         def wrong_build_request(headers):
             return "42"
@@ -1100,7 +1100,7 @@ class CommonClientServerTests:
             self.start_client()
 
     @with_server()
-    @unittest.mock.patch("websockets.server.build_response")
+    @unittest.mock.patch("websockets.asyncio_server.build_response")
     def test_server_sends_invalid_handshake_response(self, _build_response):
         def wrong_build_response(headers, key):
             return build_response(headers, "42")
@@ -1111,7 +1111,7 @@ class CommonClientServerTests:
             self.start_client()
 
     @with_server()
-    @unittest.mock.patch("websockets.client.read_response")
+    @unittest.mock.patch("websockets.asyncio_client.read_response")
     def test_server_does_not_switch_protocols(self, _read_response):
         async def wrong_read_response(stream):
             status_code, reason, headers = await read_response(stream)
@@ -1124,7 +1124,9 @@ class CommonClientServerTests:
         self.run_loop_once()
 
     @with_server()
-    @unittest.mock.patch("websockets.server.WebSocketServerProtocol.process_request")
+    @unittest.mock.patch(
+        "websockets.asyncio_server.WebSocketServerProtocol.process_request"
+    )
     def test_server_error_in_handshake(self, _process_request):
         _process_request.side_effect = Exception("process_request crashed")
 
@@ -1132,7 +1134,7 @@ class CommonClientServerTests:
             self.start_client()
 
     @with_server()
-    @unittest.mock.patch("websockets.server.WebSocketServerProtocol.send")
+    @unittest.mock.patch("websockets.asyncio_server.WebSocketServerProtocol.send")
     def test_server_handler_crashes(self, send):
         send.side_effect = ValueError("send failed")
 
@@ -1145,7 +1147,7 @@ class CommonClientServerTests:
         self.assertEqual(self.client.close_code, 1011)
 
     @with_server()
-    @unittest.mock.patch("websockets.server.WebSocketServerProtocol.close")
+    @unittest.mock.patch("websockets.asyncio_server.WebSocketServerProtocol.close")
     def test_server_close_crashes(self, close):
         close.side_effect = ValueError("close failed")
 
@@ -1220,7 +1222,9 @@ class CommonClientServerTests:
     @unittest.mock.patch(
         "websockets.server.WebSocketServerProtocol.write_http_response"
     )
-    @unittest.mock.patch("websockets.server.WebSocketServerProtocol.read_http_request")
+    @unittest.mock.patch(
+        "websockets.asyncio_server.WebSocketServerProtocol.read_http_request"
+    )
     def test_connection_error_during_opening_handshake(
         self, _read_http_request, _write_http_response
     ):
@@ -1238,7 +1242,7 @@ class CommonClientServerTests:
         _write_http_response.assert_not_called()
 
     @with_server()
-    @unittest.mock.patch("websockets.server.WebSocketServerProtocol.close")
+    @unittest.mock.patch("websockets.asyncio_server.WebSocketServerProtocol.close")
     def test_connection_error_during_closing_handshake(self, close):
         close.side_effect = ConnectionError
